# Configure Velocity classpath options for increased performance