--- a/pom.xml
+++ b/pom.xml
@@ -100,7 +100,7 @@
         <testcontainers.elasticsearch.version>1.16.2</testcontainers.elasticsearch.version>
 
         <validation.version>3.0.2</validation.version>
-        <velocity.version>2.3</velocity.version>
+        <velocity.version>2.4.1</velocity.version>
 
         <whois.main>OVERRIDE</whois.main>
     </properties>